module-1/studio/requirements.txt
langgraph
langgraph-prebuilt
langchain-core
langchain-community
langchain-google-genai
google-generativeai